Moved okta's code to the new file.
