+++ server/routes/inviteTemplate.js
@@ -1,55 +1,26 @@
 const router = require('express').Router();
-const OktaJwtVerifier = require('@okta/jwt-verifier');
 const InviteTemplate = require('../models/inviteTemplate.model');
+const okta = require('../toolset/okta');
 
-const oktaJwtVerifier = new OktaJwtVerifier({
-  issuer: `${process.env.OKTA_ISSUER}`,
-  clientId: `${process.env.OKTA_CLIENT_ID}`,
-  assertClaims: {
-    aud: 'api://default',
-  },
-});
-
-// Function from example of Okta, used to authenticate
-function authenticationRequired(req, res, next) {
-  const authHeader = req.headers.authorization || '';
-  const match = authHeader.match(/Bearer (.+)/);
-
-  if (!match) {
-    return res.status(401).end();
-  }
-
-  const accessToken = match[1];
-  const expectedAudience = 'api://default';
-
-  return oktaJwtVerifier
-      .verifyAccessToken(accessToken, expectedAudience)
-      .then((jwt) => {
-        req.jwt = jwt;
-        next();
+router
+    .route('/get/:searchQuery')
+    .get(okta.authenticationRequired, (req, res) => {
+      let regexBase = req.params.searchQuery.replace(/\s/g, '|');
+      regexBase = new RegExp(regexBase, 'gi');
+      InviteTemplate.find({
+        $or: [
+          {title: {$regex: regexBase}},
+          {instructor: {$regex: regexBase}},
+          {description: {$regex: regexBase}},
+        ],
+        $and: [{userName: req.headers.username}],
       })
-      .catch((err) => {
-        res.status(401).send(err.message);
-      });
-}
+          .then((inviteTemplate) => res.json(inviteTemplate))
+          .catch((err) => res.status(400).json('Error: ' + err));
+    });
 
-router.route('/get/:searchQuery').get(authenticationRequired, (req, res) => {
-  let regexBase = req.params.searchQuery.replace(/\s/g, '|');
-  regexBase = new RegExp(regexBase, 'gi');
-  InviteTemplate.find({
-    $or: [
-      {title: {$regex: regexBase}},
-      {instructor: {$regex: regexBase}},
-      {description: {$regex: regexBase}},
-    ],
-    $and: [{userName: req.headers.username}],
-  })
-      .then((inviteTemplate) => res.json(inviteTemplate))
-      .catch((err) => res.status(400).json('Error: ' + err));
-});
-
-// all query should have authenticationRequired like here
-router.route('/all').get(authenticationRequired, (req, res) => {
+// all query should have okta.authenticationRequired like here
+router.route('/all').get(okta.authenticationRequired, (req, res) => {
   InviteTemplate.find({
     $and: [{userName: req.headers.username}],
   })
@@ -57,7 +28,7 @@ router.route('/all').get(authenticationRequired, (req, res) => {
       .catch((err) => res.status(400).json('Error: ' + err));
 });
 
-router.route('/pending').get(authenticationRequired, (req, res) => {
+router.route('/pending').get(okta.authenticationRequired, (req, res) => {
   InviteTemplate.find({
     $and: [{userName: req.headers.username}, {sent: false}],
   })
@@ -65,7 +36,7 @@ router.route('/pending').get(authenticationRequired, (req, res) => {
       .catch((err) => res.status(400).json('Error: ' + err));
 });
 
-router.route('/sent').get(authenticationRequired, (req, res) => {
+router.route('/sent').get(okta.authenticationRequired, (req, res) => {
   InviteTemplate.find({
     $and: [{userName: req.headers.username}, {sent: true}],
   })
@@ -73,7 +44,7 @@ router.route('/sent').get(authenticationRequired, (req, res) => {
       .catch((err) => res.status(400).json('Error: ' + err));
 });
 
-router.route('/openTraining').get((req, res) => {
+router.route('/openTraining').get(okta.authenticationRequired, (req, res) => {
   InviteTemplate.find({
     $and: [{openTrainging: true}],
   })
@@ -81,7 +52,7 @@ router.route('/openTraining').get((req, res) => {
       .catch((err) => res.status(400).json('Error: ' + err));
 });
 
-router.route('/save').post((req, res) => {
+router.route('/save').post(okta.authenticationRequired, (req, res) => {
   // req.body is a template as JSON
   const template = req.body;
   const newTemplate = new InviteTemplate(template);
@@ -92,14 +63,14 @@ router.route('/save').post((req, res) => {
       .catch((err) => res.status(400).json('Error: ' + err));
 });
 
-router.route('/update/:id').post((req, res) => {
+router.route('/update/:id').post(okta.authenticationRequired, (req, res) => {
   const template = req.body;
   InviteTemplate.updateOne({_id: req.params.id}, template).then(() =>
     res.json('Template updated.'),
   );
 });
 
-router.route('/delete/:_id').post((req, res) => {
+router.route('/delete/:_id').post(okta.authenticationRequired, (req, res) => {
   // req.params => template's ID
   InviteTemplate.deleteOne(req.params)
       .then(() => res.json('Template deleted!'))
